move gdal/ogr exporters from project module to transforms/writers packages
