+++ gis4wrf/core/writers/shapefile.py
@@ -1,0 +1,8 @@
+from gis4wrf.core.util import export, ogr
+
+# TODO currently unused, remove?
+#      can be used to save domain outlines to shapefiles
+@export
+def write_shapefile(path: str, data: ogr.DataSource) -> None:
+    drv = ogr.GetDriverByName('ESRI Shapefile') # type: ogr.Driver
+    ds = drv.CopyDataSource(data, path) # type: ogr.DataSource
